edits made(add logger)

+++ demo/src/main/java/simbirSoftPractice/demo/service/implement/ItemServiceImpl.java
@@ -1,10 +1,12 @@
 package simbirSoftPractice.demo.service.implement;
 
+
+
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import simbirSoftPractice.demo.dao.entity.Company;
 import simbirSoftPractice.demo.dao.entity.Item;
-import simbirSoftPractice.demo.dao.entity.Status;
 import simbirSoftPractice.demo.dao.repository.CompanyRepository;
 import simbirSoftPractice.demo.dao.repository.ItemRepository;
 import simbirSoftPractice.demo.dto.ItemDto;
@@ -12,9 +14,13 @@ import simbirSoftPractice.demo.service.interfaces.ItemService;
 
 import java.util.List;
 
+import static org.apache.logging.log4j.LogManager.getLogger;
+
+
 @Service
 public class ItemServiceImpl implements ItemService {
 
+    private static Logger logger = getLogger(ItemServiceImpl.class);
     @Autowired
     private final ItemRepository itemRepo;
 
@@ -28,44 +34,40 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public List<Item> findAll() {
-        return (List<Item>) itemRepo.findAll();
+        List<Item> items = (List<Item>) itemRepo.findAll();
+        logger.info("successfully find all items!");
+        return items;
     }
 
     @Override
-    public Item getById(Long id) {
-        return itemRepo.findById(id).get();
+    public ResponseEntity<Item> getById(Long id) {
+        if (itemRepo.findById(id).isPresent()){
+            Item item = itemRepo.findById(id).get();
+            logger.info("successfully find item by id!");
+            return ResponseEntity.ok(item);
+        }else{
+            logger.error("EXCEPTION when delete item!");
+            return ResponseEntity.notFound().build();
+        }
+
     }
 
     @Override
     public void save(ItemDto newItem) {
-        Item item = new Item();
-        item.setName(newItem.getName());
-        item.setPrice(newItem.getPrice());
-        item.setQuantity(newItem.getQuantity());
-        Status status = new Status();
-        status.setName(newItem.getStatus());
-        item.setStatus(status);
-        Iterable<Company> companies = companyRepo.findAll();
-        companies.forEach(company -> {
-                    if (company.getName().toLowerCase().equals(newItem.getCompanyDto().toLowerCase())) {
-                        item.setCompany(company);
-                        System.out.println("-----OLD COMPANY------: " + item.toString());
-                        itemRepo.save(item);
-                        return;
-                    }
-        });
-
-                Company newCompany = new Company();
-                newCompany.setName(newItem.getCompanyDto());
-                item.setCompany(newCompany);
-                System.out.println("----NEW COMPANY----:" + item.toString());
-                itemRepo.save(item);
-                return;
-
+        itemRepo.save(newItem.itemDtoToItem(newItem));
+        logger.info("successfully save new item!");
     }
 
     @Override
-    public void deleteById(Long id) {
-        itemRepo.deleteById(id);
+    public ResponseEntity<String> deleteById(Long id) {
+        if (itemRepo.findById(id).isPresent()){
+            Item deleteItem = itemRepo.findById(id).get();
+            itemRepo.delete(deleteItem);
+            logger.info("successfully delete item!");
+            return ResponseEntity.ok("successfully delete!");
+        }else {
+            logger.error("EXCEPTION when delete item!");
+            return ResponseEntity.ok("error!!!");
+        }
     }
 }
